refactor: move defineNuxtLink to components
