+++ packages/nuxt3/src/app/components/index.ts
@@ -1,0 +1,3 @@
+// defineNuxtLink
+export { defineNuxtLink } from './nuxt-link'
+export type { NuxtLinkOptions, NuxtLinkProps } from './nuxt-link'
