--- packages/nuxt3/src/app/components/nuxt-link.mjs
@@ -1,3 +1,0 @@
-import { defineNuxtLink } from '#app'
-
-export default defineNuxtLink({ componentName: 'NuxtLink' })
