+++ packages/nuxt3/src/app/components/nuxt-link.ts
@@ -182,3 +182,5 @@ export function defineNuxtLink (options: NuxtLinkOptions) {
     }
   }) as unknown as DefineComponent<NuxtLinkProps>
 }
+
+export default defineNuxtLink({ componentName: 'NuxtLink' })
