+++ packages/nuxt3/src/app/composables/index.ts
@@ -11,5 +11,3 @@ export type { CookieOptions, CookieRef } from './cookie'
 export { useRequestHeaders } from './ssr'
 export { abortNavigation, addRouteMiddleware, defineNuxtRouteMiddleware, navigateTo, useRoute, useRouter } from './router'
 export type { AddRouteMiddlewareOptions, RouteMiddleware } from './router'
-export { defineNuxtLink } from './nuxt-link'
-export type { NuxtLinkOptions, NuxtLinkProps } from './nuxt-link'
